Fix deprecation warnings and more dtypes
Fixed pandas warnings:
- FutureWarning: The default dtype for empty Series will be 'object' instead of 'float64' in a future version. Specify a dtype explicitly to silence this warning.
- FutureWarning: Passing a set as an indexer is deprecated and will raise in a future version. Use a list instead.

+++ tests/test_spike_report.py
@@ -21,7 +21,7 @@ def _create_series(node_ids, index, name="ids"):
         return pd.Index(ids, name="times")
     if len(node_ids) == 0:
         # removing warning
-        return pd.Series(node_ids, index=_get_index(index), name=name, dtype=np.float64)
+        return pd.Series(node_ids, index=_get_index(index), name=name, dtype=np.int64)
     return pd.Series(node_ids, index=_get_index(index), name=name)
